feat(wasi): Adds WASI runtime
This is a first, working pass (with an example!) at a WASI compatible
runtime. It can retrieve logs from process, mount directories from
the host, pass args, and pass env variables.

As for testing, because of the likelihood of change as we iterate, I
have not added any unit tests, but can add them as desired. I think
the best way for testing would be to use the WASM module I dropped
in `examples/` and check that the output contains everything we passed